cxr and cxs tangential RTP frame both now working, use them to investigate solidXJfixture overlap issue, adding metadata for cross-machine plot labelling

+++ CSGOptiX/grab.sh
@@ -10,6 +10,11 @@ The remote directory to grab is configurable with envvar OPTICKS_KEYDIR_GRABBED,
 
    .opticks/geocache/DetSim0Svc_pWorld_g4live/g4ok_gltf/3dbec4dc3bdef47884fe48af781a179d/1
 
+NB to update the CSGFoundry geometry on laptop for correct labelling of volumes use::
+
+   ./cf_grab.sh 
+
+
 EOU
 }
 
@@ -40,6 +45,17 @@ printf "to                     %s \n" "$to"
 
 mkdir -p $to
 
+
+open_last(){
+    local msg="=== $FUNCNAME :"
+    local typ=${1:-jpg}
+    local last=$(ls -1rt `find ${to%/} -name "*.$typ" ` | tail -1 )
+    echo $msg typ $typ last $last
+    if [ "$(uname)" == "Darwin" ]; then
+        open $last 
+    fi 
+}
+
 if [ "$arg" == "tab" ]; then
 
     globptn="${to}cvd1/70000/cxr_overview/cam_0_tmin_0.4/cxr_overview*.jpg"
@@ -53,16 +69,23 @@ elif [ "$arg" == "tab_water" ]; then
     ${IPYTHON:-ipython} -i $(which snap.py) --  --globptn "$globptn"  $*
 
 elif [ "$arg" == "png" ]; then
+
     rsync -zarv --progress --include="*/" --include="*.png" --exclude="*" "$from" "$to"
     ls -1rt `find ${to%/} -name '*.png'`
+    open_last png
 
 elif [ "$arg" == "jpg" ]; then
+
     rsync -zarv --progress --include="*/" --include="*.jpg" --include="*.json" --exclude="*" "$from" "$to"
     ls -1rt `find ${to%/} -name '*.jpg' -o -name '*.json'`
+    open_last jpg
 
 elif [ "$arg" == "mp4" ]; then
+
     rsync -zarv --progress --include="*/" --include="*.mp4" --include="*.json" --exclude="*" "$from" "$to"
     ls -1rt `find ${to%/} -name '*.mp4' -o -name '*.json'`
+    open_last mp4
+
 
 elif [ "$arg" == "all" ]; then
     rsync -zarv --progress --include="*/" --include="*.txt" --include="*.npy" --include="*.jpg" --include="*.mp4" --include "*.json" --exclude="*" "$from" "$to"
@@ -71,20 +94,23 @@ elif [ "$arg" == "all" ]; then
     ls -1rt `find ${to%/} -name '*.jpg' -o -name '*.mp4' -o -name '*.npy'  `
 
 
-    ## write source-able script ${EXECUTABLE}_OUTPUT_DIR.sh defining correponding envvar
-    ## depending on the path of the last .npy grabbed  
-    ## This is used from cxs.sh to transparently communicate the last OUTPUT_DIR 
-    ## between nodes. 
-     
-    last_npy=$(ls -1rt `find ${to%/} -name '*.npy' ` | tail -1 )
-    last_outdir=$(dirname $last_npy)
-    script_outdir=$LOGDIR/${EXECUTABLE}_OUTPUT_DIR.sh
-    echo last_npy $last_npy 
-    echo last_outdir $last_outdir 
-    echo script_outdir $script_outdir
-    printf "export ${EXECUTABLE}_OUTPUT_DIR=$last_outdir\n" > $script_outdir
-    cat $script_outdir
+    if [ "$EXECUTABLE" == "CSGOptiXSimulateTest" -o "$EXECUTABLE" == "CSGOptiXRenderTest" ]; then  
+
+        last_npy=$(ls -1rt `find ${to%/} -name '*.npy' ` | tail -1 )
+        last_outdir=$(dirname $last_npy)
+
+        ## write source-able script ${EXECUTABLE}_OUTPUT_DIR.sh defining correponding envvar
+        ## depending on the path of the last .npy grabbed  
+        ## This is used from cxs.sh to transparently communicate the last OUTPUT_DIR 
+        ## between nodes. 
+
+        script_outdir=$LOGDIR/${EXECUTABLE}_OUTPUT_DIR.sh
+        echo last_npy $last_npy 
+        echo last_outdir $last_outdir 
+        echo script_outdir $script_outdir
+        printf "export ${EXECUTABLE}_OUTPUT_DIR=$last_outdir\n" > $script_outdir
+        cat $script_outdir
+   fi 
 
 fi 
-
